move ExecutionController::run() into ReadylogContext
it's very readylog-specific and won't necessarily be useful for other
semantics impls
